+++ experiments/browser_based_querying/src/index.tsx
@@ -2,11 +2,11 @@ import 'core-js/stable';
 import { render } from 'react-dom';
 import { createRoot } from 'react-dom/client';
 
-import App from './App';
+import HackerNews from './playgrounds/HackerNews';
 
 const mainEl = document.getElementsByTagName('main')[0];
 const root = createRoot(mainEl);
-root.render(<App />);
+root.render(<HackerNews />);
 if (process.env.NODE_ENV === 'development' && module.hot) {
-  module.hot.accept('./App', () => render(<App />, mainEl));
+  module.hot.accept('./App', () => render(<HackerNews />, mainEl));
 }
